dear

+++ src/dsxindexer/factors/float_factor.py
@@ -6,8 +6,8 @@ from dsxindexer.factors.base_factor import BaseFactor
 
 class FloatFactor(BaseFactor):
     # 设置处理标识类型
-    type_name = TokenType.INTEGER
-    
+    type_name = TokenType.FLOAT
+
     def call(self):
         # 处理浮点数
         self.parser.eat(self.type_name)
